refactor: 구글 맵을 로드하는 컨테이너 컴포넌트 분리 및 코드 리팩터링
